+++ src/generations/features/herald.feature.ts
@@ -1,10 +1,10 @@
 import { communicatorsConfigStub } from "../stubs";
-import { FeatureDefinition } from "./types";
+import { FeatureDefinition, INSTALLED_WARLOCK_VERSION } from "./types";
 
 export const heraldFeature: FeatureDefinition = {
   description: "Installs herald for message broker (Herald Package)",
   dependencies: {
-    "@warlock.js/herald": "~4.0.0",
+    "@warlock.js/herald": INSTALLED_WARLOCK_VERSION,
     amqplib: "^0.10.0",
   },
   devDependencies: {